alert_notification/dashboard: fix compatibility with grafana 5.0
This strips the new `uid` field, soon to be replacing `slug`, to
fix #15.

Additionally, it adds a check for a null 200 OK response for
alert notifications by checking both the content of the response
and the presence of errors. This maybe is something we'd want to change
in the client library but seems like an easy approach to getting
the tests passing.