Simplify test target
No need to call 'go test' with -i first. The -i flag installs packages
that are dependencies of the test, but does not run the tests. This
isn't useful since dependencies are vendored.

This will also fail if the user executing the command does not have
write permission on the Go installation.