Rename grafana_synthetic_monitoring_probe examples
simple.tf and complex.tf didn't make sense because they are essentially
the same. Instead resource.tf and resource_update.tf since the reason
there is two is to check that update works properly. The template is
also deleted in favor of the tfplugindocs default which will render
resource.tf automatically.