Fix folder ID documentation
By adding the ID as a computed attribute it will show as a read-only
attribute instead of the misleading comment that it might be possible
to assign a value. This documentation seem to be generated
automatically by the tfplugindocs tool, so we might need to either fix
that or changing the way we generate the documentation.

Closes #262.

Signed-off-by: Leandro López (inkel) <inkel.ar@gmail.com>