Allow configuring a report using a dashboard UID
The whole structure of the reports API has changed. It now allows multiple dashboards.
However, I'm not implementing this in this PR. I'm just changing the API to allow using UIDs because they are now the standard way to identify dashboards.

If we want to support multiple dashboards, it should be a completely different attribute anyways.