Add support for Machine Learning holidays
Note that holidays can be created and added to jobs, but not vice
versa, despite the API supporting it. This is because the API
would return an updated value for the holiday's `jobs` field if we
exposed it via Terraform, and we have no way of updating that from
the job resource's CreateContext (as far as I'm aware).

NOTE: this currently depends on an unreleased version of the
machine-learning-go-client repo; once
https://github.com/grafana/machine-learning-go-client/pull/5 is merged,
I'll adjust this PR and mark it ready for review.