Fix "create folder from editor key" test (#779)
* Fix "create folder from editor key" test
Closes https://github.com/grafana/terraform-provider-grafana/issues/773
The issue was that you can't create a new `provider` instance from tests. You need to use the factories functions in the test framework
The tests were failing in a random fashion because it seems like the test framework was sometimes using one provider to delete the api key, sometimes using the other

To fix the test, I instead create an API key and inject it in envvars at the test level (not in the Terraform config)
I also added a test for Viewers, it's a sanity check to make sure the auth is inject correctly. Using a Viewer api key causes an error

* Add comment

* fix lint