Convert `org_id` references to TypeString (#791)
This is strictly for the crossplane provider. It doesn't affect the Terraform provider
The Crossplane provider only works with string references, because IDs are always strings in Terraform
Terraform, however, doesn't care and will happily convert between int and string, which is why there is no migration needed from this PR. You can even go back to the int version and it will convert it right back to int.