Linkchecking: Fix all redirects (#815)
I noticed that lots of links were broken in a weird way. They are marked as `latest` but redirect to v8.4 docs, which is extremely annoying
The naming scheme changed with the new docs website so I assume that it fails to find the new reference and badly redirects
I used https://github.com/linkchecker/linkchecker/pull/619 to find all redirects. I should help to get that merged. Linkchecker is still active but not that PR