Delete gracefully if user no longer exist in grafana
When applying a terraform plan that was created with `-refresh=false` the
detection of resources that no longer exist in grafana does not work,
as the resources are not read at all.

Thus on apply, a deletion of a user that no longer exists in grafana fails with:
```
Error: status: 404, body: {"message":"user not found"}
```

This change modifies `DeleteUser` so that it catches this case and does
not return an error.