SM: rename checks to assertions (#1111)
Synthetic Monitoring has a resource called `checks`, and for the
specific case of MultiHTTP checks, each request can have zero or more
"assertions", which in the code are called "checks" because they check
that something is true. Let's not expose that to the user, as it will
create confunsion. The UI calls these things "assertions", so we can use
that term consistenly.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>