Alerting Contact Points: Refer by name (#1247)
* Alerting Contact Points: Refer by name
Background:
Currently, the contact point's ID is a collection of notifier UIDs. However, in Grafana (and in the resource's import ref), it's referred to by name

This PR:
Change the Terraform ID from semi-colon separated UIDs to the contact point name. This will be easier to test, and this will allow simplifying the code in the next major version (removing the code fetching notifiers by UID)
Since the UID reading logic is moved to the read function, it also allows importing contact points in Crossplane where the import function is not used

* ForceNew on the name because it's now the primary ID