Force recomputation of computed values for Grafana Cloud stacks on slug update (#1269)
* Force recomputation of computed values for Grafana Cloud stacks on slug update

Names of some attributes of the Grafana Cloud stack depend on the slug
of the stack. While the current state of the provider does update these
attributes on the resource in the state, if these attributes are
referenced in an output clause, the absence of a change in the generated
plan renders the change unnoticed and stale information is outputted on
a first apply. By using https://pkg.go.dev/github.com/hashicorp/terraform-plugin-sdk/helper/customdiff
it's possible to mark these values to be computed again if the slug
changes, thus being reflected as changes in the plan and generated an
up to date output.

Fixes #1191

* Adjust code formatting with go fmt

---------

Co-authored-by: Stefano Boriero <stefano.boriero@ocado.com>