POC: Standardize and verify ID formats
Currently, resource ID formats are all over the place.
The separator is mostly `:`, but sometimes it's `_`, `/` or `;`

In this PR, I create a helper struct to manage IDs:
- It defines a default separator (with retrocompatibility with the old formats
- It provides validation for the number of fields provided. Currently, it panics in most resources if you provide not enough fields
- It does auto-generation of the `import.sh` files

If this looks good, I can expand this to all resources and then validate that the ID is set this way for all resources. Providing consistent import and ID management for all resources!