Code Generation: Initial tests (#1583)
Implements a full integration test using the dashboard example file:
This test applies the dashboard shown in the TF docs and tests that the expected files are there after generating the TF config from the instance

To make this test pass reliably, I had to also make sure the files being generated are properly sorted. This is also tested

Note: The generated files are not optimal. This PR is not meant to modify behaviour