putting in a check for key/value for labels
Each label should have both a "key" and a "value" assigned. We can't
force Terraform's schema to do that using TypeMap.

What we can do, though, is check they exist and skip the label then.
It'd be better to warn the user or error on it, but this is better than
a panic.