changing how id is populated
a label is unique by the key/value pair, so the id for that should
reflect that fact.

We could potentially do a hash on this as well, but this is probably ok
for now. As colons aren't legal characters, this should remain unique
without issues like the key or value including it.