feat: Oncall integration labels (#2125)
OnCall integrations created in terraform will have an optional field called labels that is a zero or more list of label objects. Each label object requires a key and a value. Missing either and it will skip creating a label.

For ease and reuse, we can create a label datasource for oncall to pass in.