feat: add Backstage catalog files (#2228)
The catalog files are generated from internal Google sheet, the intention is to
replace this sheet with entries in Backstage (EngHub) and keep it in
sync with the repository.

This commit adds the generated catalog files and a GHA to validate
whether the catalog files are in sync with the provider resources.