fix: SM client reports unknown version if provider version is empty
Some API logs indicated that the requests were coming from the TF
provider but the version was the default version of the underlying
API go client. This would only occur in cases where the provider set
an empty string for the custom version, causing the go client's
version to override the empty value.

This explicitly provides 'unknown' for such cases so that we don't
have an explicitly wrong version being reported for Terraform.